+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugfixes
 
 * Add `for`/`id` attributes to filter labels and inputs for accessibility (#91)
+* Fix `match_mode` option being ignored for multi-select filters on relationship fields
 
 ## v0.9.0-beta.4 (2026-01-26)